+++ relay-ourlogs/src/ourlog.rs
@@ -52,7 +52,6 @@ pub fn otel_to_sentry_log(otel_log: OtelLog, received_at: DateTime<Utc>) -> Resu
 
     let mut attribute_data = Attributes::default();
 
-    attribute_data.insert("sentry.severity_text".to_owned(), severity_text.clone());
     attribute_data.insert("sentry.severity_number".to_owned(), severity_number as i64);
     attribute_data.insert(
         "sentry.timestamp_nanos".to_owned(),
@@ -63,9 +62,6 @@ pub fn otel_to_sentry_log(otel_log: OtelLog, received_at: DateTime<Utc>) -> Resu
         "sentry.observed_timestamp_nanos".to_owned(),
         received_at_nanos.to_string(),
     );
-    attribute_data.insert("sentry.trace_flags".to_owned(), 0);
-    attribute_data.insert("sentry.body".to_owned(), body.clone());
-    attribute_data.insert("sentry.span_id".to_owned(), span_id.to_string());
 
     for attribute in attributes.into_iter() {
         if let Some(value) = attribute.value.and_then(|v| v.value) {
@@ -130,18 +126,6 @@ pub fn ourlog_merge_otel(ourlog: &mut Annotated<OurLog>, received_at: DateTime<U
         .unwrap_or_else(|| UnixTimestamp::now().as_nanos() as i64);
 
     attributes.insert(
-        "sentry.severity_text".to_owned(),
-        ourlog_value
-            .level
-            .value()
-            .map(|level| level.to_string())
-            .unwrap_or_else(|| "info".to_owned()),
-    );
-    attributes.insert(
-        "sentry.severity_number".to_owned(),
-        level_to_otel_severity_number(ourlog_value.level.value().cloned()),
-    );
-    attributes.insert(
         "sentry.timestamp_nanos".to_owned(),
         timestamp_nanos.to_string(),
     );
@@ -150,29 +134,6 @@ pub fn ourlog_merge_otel(ourlog: &mut Annotated<OurLog>, received_at: DateTime<U
         "sentry.observed_timestamp_nanos".to_owned(),
         received_at_nanos.to_string(),
     );
-    attributes.insert("sentry.trace_flags".to_owned(), 0);
-    attributes.insert(
-        "sentry.body".to_owned(),
-        ourlog_value.body.value().cloned().unwrap_or_default(),
-    );
-
-    if let Some(span_id) = ourlog_value.span_id.value() {
-        attributes.insert("sentry.span_id".to_owned(), span_id.to_string());
-    }
-}
-
-fn level_to_otel_severity_number(level: Option<OurLogLevel>) -> i64 {
-    match level {
-        Some(OurLogLevel::Trace) => 1,
-        Some(OurLogLevel::Debug) => 5,
-        Some(OurLogLevel::Info) => 9,
-        Some(OurLogLevel::Warn) => 13,
-        Some(OurLogLevel::Error) => 17,
-        Some(OurLogLevel::Fatal) => 21,
-        // 0 is the default value.
-        // https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/68e1d6cd94bfca9bdf725327d4221f97ce0e0564/pkg/stanza/docs/types/severity.md
-        _ => 0,
-    }
 }
 
 #[cfg(test)]
@@ -441,25 +402,9 @@ mod tests {
               "type": "string",
               "value": "9"
             },
-            "sentry.body": {
-              "type": "string",
-              "value": "Example log record"
-            },
             "sentry.observed_timestamp_nanos": {
               "type": "string",
               "value": "946684800000000000"
-            },
-            "sentry.severity_number": {
-              "type": "integer",
-              "value": 9
-            },
-            "sentry.severity_text": {
-              "type": "string",
-              "value": "info"
-            },
-            "sentry.span_id": {
-              "type": "string",
-              "value": "eee19b7ec3c1b174"
             },
             "sentry.timestamp_nanos": {
               "type": "string",
@@ -468,47 +413,10 @@ mod tests {
             "sentry.timestamp_precise": {
               "type": "integer",
               "value": 946684800000000000
-            },
-            "sentry.trace_flags": {
-              "type": "integer",
-              "value": 0
             }
           }
         }
         "###);
-    }
-
-    #[test]
-    fn ourlog_merge_otel_log_with_unknown_severity_number() {
-        let json = r#"{
-            "timestamp": 946684800.0,
-            "level": "abc",
-            "trace_id": "5B8EFFF798038103D269B633813FC60C",
-            "span_id": "EEE19B7EC3C1B174",
-            "body": "Example log record",
-            "attributes": {
-                "foo": {
-                    "value": "9",
-                    "type": "string"
-                }
-            }
-        }"#;
-
-        let mut data = Annotated::<OurLog>::from_json(json).unwrap();
-        ourlog_merge_otel(
-            &mut data,
-            DateTime::from_timestamp_nanos(946684800000000000),
-        );
-        assert_eq!(
-            data.value()
-                .unwrap()
-                .attributes
-                .value()
-                .unwrap()
-                .get_attribute("sentry.severity_number")
-                .unwrap(),
-            &Attribute::new(AttributeType::Integer, Value::I64(0)),
-        );
     }
 
     #[test]
@@ -538,21 +446,9 @@ mod tests {
               "type": "string",
               "value": "9"
             },
-            "sentry.body": {
-              "type": "string",
-              "value": "somebody"
-            },
             "sentry.observed_timestamp_nanos": {
               "type": "string",
               "value": "946684800000000000"
-            },
-            "sentry.severity_number": {
-              "type": "integer",
-              "value": 0
-            },
-            "sentry.severity_text": {
-              "type": "string",
-              "value": "info"
             },
             "sentry.timestamp_nanos": {
               "type": "string",
@@ -561,10 +457,6 @@ mod tests {
             "sentry.timestamp_precise": {
               "type": "integer",
               "value": 1638144000000000000
-            },
-            "sentry.trace_flags": {
-              "type": "integer",
-              "value": 0
             }
           }
         }