+++ relay-server/src/services/store.rs
@@ -109,7 +109,7 @@ pub struct StoreMetrics {
 
 /// Publishes a log item to Sentry core application through Kafka.
 #[derive(Debug)]
-pub struct StoreLog {
+pub struct StoreTraceItem {
     /// The final trace item which will be produced to Kafka.
     pub trace_item: TraceItem,
     /// Outcomes to be emitted when successfully producing the item to Kafka.
@@ -119,7 +119,7 @@ pub struct StoreLog {
     pub quantities: Quantities,
 }
 
-impl Counted for StoreLog {
+impl Counted for StoreTraceItem {
     fn quantities(&self) -> Quantities {
         self.quantities.clone()
     }
@@ -134,15 +134,15 @@ pub enum Store {
     /// An envelope containing a mixture of items.
     ///
     /// Note: Some envelope items are not supported to be submitted at all or through an envelope,
-    /// for example logs must be submitted via [`Self::Log`] instead.
+    /// for example logs must be submitted via [`Self::TraceItem`] instead.
     ///
     /// Long term this variant is going to be replaced with fully typed variants of items which can
     /// be stored instead.
     Envelope(StoreEnvelope),
     /// Aggregated generic metrics.
     Metrics(StoreMetrics),
-    /// A singular log item.
-    Log(Managed<StoreLog>),
+    /// A singular [`TraceItem`].
+    TraceItem(Managed<StoreTraceItem>),
 }
 
 impl Store {
@@ -151,7 +151,7 @@ impl Store {
         match self {
             Store::Envelope(_) => "envelope",
             Store::Metrics(_) => "metrics",
-            Store::Log(_) => "log",
+            Store::TraceItem(_) => "log",
         }
     }
 }
@@ -174,11 +174,11 @@ impl FromMessage<StoreMetrics> for Store {
     }
 }
 
-impl FromMessage<Managed<StoreLog>> for Store {
+impl FromMessage<Managed<StoreTraceItem>> for Store {
     type Response = NoResponse;
 
-    fn from_message(message: Managed<StoreLog>, _: ()) -> Self {
-        Self::Log(message)
+    fn from_message(message: Managed<StoreTraceItem>, _: ()) -> Self {
+        Self::TraceItem(message)
     }
 }
 
@@ -217,7 +217,7 @@ impl StoreService {
             match message {
                 Store::Envelope(message) => self.handle_store_envelope(message),
                 Store::Metrics(message) => self.handle_store_metrics(message),
-                Store::Log(message) => self.handle_store_log(message),
+                Store::TraceItem(message) => self.handle_store_trace_item(message),
             }
         })
     }
@@ -343,7 +343,17 @@ impl StoreService {
                 ItemType::Span => {
                     self.produce_span(scoping, received_at, event_id, retention, item)?
                 }
-                ItemType::Log => self.produce_log(scoping, received_at, retention, item)?,
+                ty @ ItemType::Log => {
+                    debug_assert!(
+                        false,
+                        "received {ty} through an envelope, \
+                        this item must be submitted via a specific store message instead"
+                    );
+                    relay_log::error!(
+                        tags.project_key = %scoping.project_key,
+                        "StoreService received unsupported item type '{ty}' in envelope"
+                    );
+                }
                 ItemType::ProfileChunk => self.produce_profile_chunk(
                     scoping.organization_id,
                     scoping.project_id,
@@ -515,25 +525,23 @@ impl StoreService {
         }
     }
 
-    fn handle_store_log(&self, message: Managed<StoreLog>) {
+    fn handle_store_trace_item(&self, message: Managed<StoreTraceItem>) {
         let scoping = message.scoping();
         let received_at = message.received_at();
 
-        let quantities = message.try_accept(|log| {
+        let quantities = message.try_accept(|item| {
+            let item_type = item.trace_item.item_type();
             let message = KafkaMessage::Item {
                 headers: BTreeMap::from([
                     ("project_id".to_owned(), scoping.project_id.to_string()),
-                    (
-                        "item_type".to_owned(),
-                        (TraceItemType::Log as i32).to_string(),
-                    ),
+                    ("item_type".to_owned(), (item_type as i32).to_string()),
                 ]),
-                message: log.trace_item,
-                item_type: TraceItemType::Log,
+                message: item.trace_item,
+                item_type,
             };
 
             self.produce(KafkaTopic::Items, message)
-                .map(|()| log.quantities)
+                .map(|()| item.quantities)
         });
 
         // Accepted outcomes when items have been successfully produced to rdkafka.
@@ -1281,142 +1289,6 @@ impl StoreService {
         Ok(())
     }
 
-    fn produce_log(
-        &self,
-        scoping: Scoping,
-        received_at: DateTime<Utc>,
-        retention_days: u16,
-        item: &Item,
-    ) -> Result<(), StoreError> {
-        relay_log::trace!("Producing log");
-
-        let payload = item.payload();
-        let d = &mut Deserializer::from_slice(&payload);
-        let logs: LogKafkaMessages = match serde_path_to_error::deserialize(d) {
-            Ok(logs) => logs,
-            Err(error) => {
-                relay_log::error!(
-                    error = &error as &dyn std::error::Error,
-                    "failed to parse log"
-                );
-                self.outcome_aggregator.send(TrackOutcome {
-                    category: DataCategory::LogItem,
-                    event_id: None,
-                    outcome: Outcome::Invalid(DiscardReason::InvalidLog),
-                    quantity: 1,
-                    remote_addr: None,
-                    scoping,
-                    timestamp: received_at,
-                });
-                self.outcome_aggregator.send(TrackOutcome {
-                    category: DataCategory::LogByte,
-                    event_id: None,
-                    outcome: Outcome::Invalid(DiscardReason::InvalidLog),
-                    quantity: payload.len() as u32,
-                    remote_addr: None,
-                    scoping,
-                    timestamp: received_at,
-                });
-                return Ok(());
-            }
-        };
-
-        let num_logs = logs.items.len() as u32;
-        for log in logs.items {
-            let timestamp_seconds = log.timestamp as i64;
-            let timestamp_nanos = (log.timestamp.fract() * 1e9) as u32;
-            let item_id = u128::from_be_bytes(
-                *Uuid::new_v7(uuid::Timestamp::from_unix(
-                    uuid::NoContext,
-                    timestamp_seconds as u64,
-                    timestamp_nanos,
-                ))
-                .as_bytes(),
-            )
-            .to_le_bytes()
-            .to_vec();
-            let mut trace_item = TraceItem {
-                item_type: TraceItemType::Log.into(),
-                organization_id: scoping.organization_id.value(),
-                project_id: scoping.project_id.value(),
-                received: Some(Timestamp {
-                    seconds: safe_timestamp(received_at) as i64,
-                    nanos: 0,
-                }),
-                retention_days: retention_days.into(),
-                timestamp: Some(Timestamp {
-                    seconds: timestamp_seconds,
-                    nanos: 0,
-                }),
-                trace_id: log.trace_id.to_string(),
-                item_id,
-                attributes: Default::default(),
-                client_sample_rate: 1.0,
-                server_sample_rate: 1.0,
-            };
-
-            for (name, attribute) in log.attributes.unwrap_or_default() {
-                if let Some(attribute_value) = attribute {
-                    if let Some(v) = attribute_value.value {
-                        let any_value = match v {
-                            LogAttributeValue::String(value) => AnyValue {
-                                value: Some(Value::StringValue(value)),
-                            },
-                            LogAttributeValue::Int(value) => AnyValue {
-                                value: Some(Value::IntValue(value)),
-                            },
-                            LogAttributeValue::Bool(value) => AnyValue {
-                                value: Some(Value::BoolValue(value)),
-                            },
-                            LogAttributeValue::Double(value) => AnyValue {
-                                value: Some(Value::DoubleValue(value)),
-                            },
-                            LogAttributeValue::Unknown(_) => continue,
-                        };
-
-                        trace_item.attributes.insert(name.into(), any_value);
-                    }
-                }
-            }
-
-            let message = KafkaMessage::Item {
-                item_type: TraceItemType::Log,
-                headers: BTreeMap::from([
-                    ("project_id".to_owned(), scoping.project_id.to_string()),
-                    (
-                        "item_type".to_owned(),
-                        (TraceItemType::Log as i32).to_string(),
-                    ),
-                ]),
-                message: trace_item,
-            };
-
-            self.produce(KafkaTopic::Items, message)?;
-        }
-
-        // We need to track the count and bytes separately for possible rate limits and quotas on both counts and bytes.
-        self.outcome_aggregator.send(TrackOutcome {
-            category: DataCategory::LogItem,
-            event_id: None,
-            outcome: Outcome::Accepted,
-            quantity: num_logs,
-            remote_addr: None,
-            scoping,
-            timestamp: received_at,
-        });
-        self.outcome_aggregator.send(TrackOutcome {
-            category: DataCategory::LogByte,
-            event_id: None,
-            outcome: Outcome::Accepted,
-            quantity: payload.len() as u32,
-            remote_addr: None,
-            scoping,
-            timestamp: received_at,
-        });
-
-        Ok(())
-    }
-
     fn produce_profile_chunk(
         &self,
         organization_id: OrganizationId,
@@ -1898,44 +1770,6 @@ impl SpanKafkaMessage<'_> {
     }
 }
 
-#[derive(Clone, Debug, Deserialize)]
-#[serde(tag = "type", content = "value")]
-enum LogAttributeValue {
-    #[serde(rename = "string")]
-    String(String),
-    #[serde(rename = "boolean")]
-    Bool(bool),
-    #[serde(rename = "integer")]
-    Int(i64),
-    #[serde(rename = "double")]
-    Double(f64),
-    #[serde(rename = "unknown")]
-    Unknown(()),
-}
-
-/// This is a temporary struct to convert the old attribute format to the new one.
-#[derive(Debug, Deserialize)]
-#[allow(dead_code)]
-struct LogAttribute {
-    #[serde(flatten)]
-    value: Option<LogAttributeValue>,
-}
-
-#[derive(Debug, Deserialize)]
-struct LogKafkaMessages<'a> {
-    #[serde(borrow)]
-    items: Vec<LogKafkaMessage<'a>>,
-}
-
-#[derive(Debug, Deserialize)]
-struct LogKafkaMessage<'a> {
-    trace_id: EventId,
-    #[serde(default)]
-    timestamp: f64,
-    #[serde(borrow, default)]
-    attributes: Option<BTreeMap<&'a str, Option<LogAttribute>>>,
-}
-
 fn none_or_empty_object(value: &Option<&RawValue>) -> bool {
     match value {
         None => true,
@@ -2024,8 +1858,8 @@ impl Message for KafkaMessage<'_> {
         }
     }
 
-    /// Returns the partitioning key for this kafka message determining.
-    fn key(&self) -> Option<[u8; 16]> {
+    /// Returns the partitioning key for this Kafka message determining.
+    fn key(&self) -> Option<relay_kafka::Key> {
         match self {
             Self::Event(message) => Some(message.event_id.0),
             Self::Attachment(message) => Some(message.event_id.0),
@@ -2040,40 +1874,19 @@ impl Message for KafkaMessage<'_> {
             // recieve the routing_key_hint form their envelopes.
             Self::CheckIn(message) => message.routing_key_hint,
 
-            // Generate a new random routing key, instead of defaulting to `rdkafka` behaviour
-            // for no routing key.
-            //
-            // This results in significantly more work for Kafka, but we've seen that the metrics
-            // indexer consumer in Sentry, cannot deal with this load shape.
-            // Until the metric indexer is updated, we still need to assign random keys here.
-            Self::Metric { .. } => Some(Uuid::new_v4()),
-
             // Random partitioning
-            Self::Profile(_)
-            | Self::ReplayRecordingNotChunked(_)
-            | Self::ProfileChunk(_)
-            | Self::Item { .. } => None,
+            _ => None,
         }
         .filter(|uuid| !uuid.is_nil())
-        .map(|uuid| uuid.into_bytes())
+        .map(|uuid| uuid.as_u128())
     }
 
     fn headers(&self) -> Option<&BTreeMap<String, String>> {
         match &self {
             KafkaMessage::Metric { headers, .. }
             | KafkaMessage::Span { headers, .. }
-            | KafkaMessage::Item { headers, .. } => {
-                if !headers.is_empty() {
-                    return Some(headers);
-                }
-                None
-            }
-            KafkaMessage::Profile(profile) => {
-                if !profile.headers.is_empty() {
-                    return Some(&profile.headers);
-                }
-                None
-            }
+            | KafkaMessage::Item { headers, .. }
+            | KafkaMessage::Profile(ProfileKafkaMessage { headers, .. }) => Some(headers),
             _ => None,
         }
     }
@@ -2143,7 +1956,7 @@ mod tests {
         for topic in [KafkaTopic::Outcomes, KafkaTopic::OutcomesBilling] {
             let res = producer
                 .client
-                .send(topic, Some(*b"0123456789abcdef"), None, "foo", b"");
+                .send(topic, Some(0x0123456789abcdef), None, "foo", b"");
 
             assert!(matches!(res, Err(ClientError::InvalidTopicName)));
         }